feat: Adição da nova coluna de escolaridade na tabela Dim_Pessoas. - Criação das novas tabelas fatos de beneficios, lógica de idade de dependentes e titulares - Novas classificações como Conjuge e Filhos(a), tabela de valores do plano de sáude compativel com a ANS - Calculo mensal atual. - fixes #11 - fixes #12

+++ iceberg_rh_dbt/models/staging/stg_cargos.sql
@@ -11,10 +11,10 @@ renamed_and_casted AS (
         departamento AS nome_departamento,
         nome_cargo AS nome_cargo,
         nivel_hierarquico AS nivel_senioridade,
-        CAST(faixa_80_min AS FLOAT64) AS faixa_80_min,
-        CAST(faixa_100_mid AS FLOAT64) AS faixa_100_mid,
-        CAST(faixa_130_max AS FLOAT64) AS faixa_130_max,
-        CAST(peso_contratacao AS FLOAT64) AS peso_contratacao
+        CAST(faixa_80_min AS FLOAT) AS faixa_80_min,
+        CAST(faixa_100_mid AS FLOAT) AS faixa_100_mid,
+        CAST(faixa_130_max AS FLOAT) AS faixa_130_max,
+        CAST(peso_contratacao AS FLOAT) AS peso_contratacao
     FROM source
 )
 SELECT *
