+++ iceberg_rh_dbt/models/staging/stg_estrutura.sql
@@ -6,15 +6,15 @@ WITH source AS (
 ),
 renamed_and_casted AS (
     SELECT
-      CAST(id_area AS INT64) AS sk_area,
+      CAST(id_area AS INT) AS sk_area,
       diretoria,
       departamento,
       centro_de_custo,
       estado_sigla,
       cidade_escritorio,
-      CAST(lat AS FLOAT64) AS latitude,
-      CAST(lon AS FLOAT64) AS longitude,
-      CAST(probabilidade_alocacao AS FLOAT64) AS probabilidade_alocacao
+      CAST(lat AS FLOAT) AS latitude,
+      CAST(lon AS FLOAT) AS longitude,
+      CAST(probabilidade_alocacao AS FLOAT) AS probabilidade_alocacao
     FROM source
 )
 SELECT *
