+++ iceberg_rh_dbt/models/staging/stg_fato_contratos.sql
@@ -6,16 +6,16 @@ WITH source AS (
 ),
 renamed_and_casted AS (
     SELECT
-        CAST(id_contrato AS INT64) AS sk_contrato,
-        CAST(id_pessoa AS INT64) AS sk_pessoa,
-        CAST(id_area AS INT64) AS sk_area,
-        CAST(id_cargo AS INT64) AS sk_cargo,
+        CAST(id_contrato AS INT) AS sk_contrato,
+        CAST(id_pessoa AS INT) AS sk_pessoa,
+        CAST(id_area AS INT) AS sk_area,
+        CAST(id_cargo AS INT) AS sk_cargo,
         tipo_contrato,
         modelo_trabalho,
         CAST(data_admissao AS DATE) AS data_admissao,
         CAST(data_demissao AS DATE) AS data_demissao,
         status AS status_contrato,
-        CAST(salario AS FLOAT64) AS salario_base
+        CAST(salario AS DOUBLE) AS salario_base
     FROM source
 )
 SELECT *
